+++ sdwan_policy_objects.tf
@@ -256,3 +256,12 @@ resource "sdwan_policy_object_security_protocol_list" "policy_object_security_pr
     protocol_name = protocol
   }]
 }
+
+resource "sdwan_policy_object_security_url_allow_list" "policy_object_security_url_allow_list" {
+  for_each           = { for p in try(local.feature_profiles.policy_object_profile.security_url_allow_lists, {}) : p.name => p }
+  name               = each.value.name
+  feature_profile_id = sdwan_policy_object_feature_profile.policy_object_feature_profile[0].id
+  entries = [for e in try(each.value.urls, []) : {
+    pattern = e
+  }]
+}