+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - add support for policy object security local application list
 - add support for policy object security port list
 - add support for policy object security protocol list
+- add support for policy object security url allow list
 - add support for `cloud_qos` and `cloud_qos_service_side` in `sdwan_localized_policy`
 
 ## 1.3.0